moving trmm_ files from exp01_gemm.py to exp07_properties_multiplication.py
